# Kdesktop 1531 optimize io helper directory iterator