### PR Description

`DirectoryIterator::next` was previously inefficient due to the creation of two iterators per call, just to obtain the `end` value. This is unnecessary, as we can retrieve the `end` value from the already constructed iterator.

Additionally, we frequently convert `std::error_code` to `ioError`, even when it’s not needed. Removing these redundant conversions simplifies the code and avoids unnecessary overhead.

These optimizations provide **significant performance improvements**:

#### Benchmark Results:

**Case 1: Recursive traversal of a folder containing 100 subfolders, each with 1,000 files and 1,000 empty directories (200,100 items)**  
- Before: **38.078s**  
- After: **6.959s**

**Case 2: Recursive traversal of a folder containing 100 subfolders, each with 2,000 files (200,100 items)**  
- Before: **20s**  
- After: **610ms**

### Summary

- No more spurrious Iterator creation.
- Removed unnecessary `ioError` conversions.
- **Up to 40x faster** in common scenarios.
- File traversal time is now almost negligible.
- As typical drives contain more files than directories, real-world improvements will generally align with the faster case.

This change significantly improves the performance of directory traversal operations, especially on large file systems.